// packages/contract-helpers/src/staking-contract/typechain/IAaveStakingHelper__factory.ts
/* Autogenerated file. Do not edit manually. */
/* eslint-disable */

import { Contract, Signer } from 'ethers';
import { Provider } from '@ethersproject/providers';

import type { IAaveStakingHelper } from './IAaveStakingHelper';

export class IAaveStakingHelper__factory {
  static connect(
    address: string,
    signerOrProvider: Signer | Provider,
  ): IAaveStakingHelper {
    return new Contract(address, _abi, signerOrProvider) as IAaveStakingHelper;
  }
}

const _abi = [
  {
    inputs: [
      {
        internalType: 'address',
        name: 'user',
        type: 'address',
      },
      {
        internalType: 'uint256',
        name: 'amount',
        type: 'uint256',
      },
      {
        internalType: 'uint8',
        name: 'v',
        type: 'uint8',
      },
      {
        internalType: 'bytes32',
        name: 'r',
        type: 'bytes32',
      },
      {
        internalType: 'bytes32',
        name: 's',
        type: 'bytes32',
      },
    ],
    name: 'stake',
    outputs: [],
    stateMutability: 'nonpayable',
    type: 'function',
  },
];